Move initial admin account logic to common helper
