+++ pkg/install/install.go
@@ -19,8 +19,6 @@ package install
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"net/url"
@@ -386,7 +384,7 @@ func (o *Install) provisionEverest(ctx context.Context, v *goversion.Version) er
 		if err := o.kubeClient.CreateRSAKeyPair(ctx); err != nil {
 			return err
 		}
-		if err := o.resetEverestAdminPassword(ctx); err != nil {
+		if err := common.CreateInitialAdminAccount(ctx, o.kubeClient.Accounts()); err != nil {
 			return err
 		}
 	} else {
@@ -748,24 +746,4 @@ func validateRFC1035(s string) error {
 	}
 
 	return nil
-}
-
-func (o *Install) resetEverestAdminPassword(ctx context.Context) error {
-	o.l.Info("Resetting admin password")
-	pass, err := generateRandomPassword()
-	if err != nil {
-		return errors.Join(err, errors.New("could not generate random password"))
-	}
-	if err := o.kubeClient.Accounts().SetPassword(ctx, common.EverestAdminUser, pass, false); err != nil {
-		return err
-	}
-	return nil
-}
-
-func generateRandomPassword() (string, error) {
-	b := make([]byte, 32)
-	if _, err := rand.Read(b); err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(b), nil
 }
